+++ ase/calculators/vasp/create_input.py
@@ -1,4 +1,3 @@
-from __future__ import print_function
 # Copyright (C) 2008 CSC - Scientific Computing Ltd.
 """This module defines an ASE interface to VASP.
 
@@ -625,6 +624,9 @@ bool_keys = [
     'qifcg',      # Use CG instead of quickmin (instanton)
     'qdo_ins',    # Find instanton
     'qdo_pre',    # Calculate prefactor (instanton)
+    # The next keyword pertains to the periodic NBO code of JR Schmidt's group
+    # at UW-Madison (https://github.com/jrschmidt2/periodic-NBO)
+    'lnbo',    # Enable NBO analysis
 ]
 
 list_int_keys = [