+++ ase/gui/add.py
@@ -1,5 +1,4 @@
 # encoding: utf-8
-from __future__ import unicode_literals
 
 import os
 import numpy as np
@@ -31,7 +30,11 @@ class AddAtoms:
             # Load the file immediately, so we can warn now in case of error
             self.readfile(filename, format=chooser.format)
 
-        self.entry = ui.Entry('', callback=self.add)
+        if self.gui.images.selected.any():
+            default = '(current selection)'
+        else:
+            default = ''
+        self.entry = ui.Entry(default, callback=self.add)
         win.add([_('Add:'), self.entry,
                  ui.Button(_('File ...'), callback=choose_file)])
 
@@ -78,6 +81,12 @@ class AddAtoms:
 
     def get_atoms(self):
         val = self.entry.value
+
+        if val == '(current selection)':
+            selection = self.gui.images.selected.copy()
+            if selection.any():
+                atoms = self.gui.atoms.copy()
+                return atoms[selection[:len(self.gui.atoms)]]
 
         if val in atomic_numbers:  # Note: This means val is a symbol!
             return Atoms(val)