+++ ase/test/vasp/vasp_xml.py
@@ -5,9 +5,7 @@ environment variables
 
 """
 
-import unittest
 import numpy as np
-import sys
 
 from ase.test.vasp import installed
 from ase import Atoms
@@ -16,9 +14,6 @@ from ase.io import read
 
 
 def main():
-    if sys.version_info < (2, 7):
-        raise unittest.SkipTest('read_xml requires Python version 2.7 or greater')
-
     assert installed()
 
     # simple test calculation of CO molecule